third_party/WebKit/LayoutTests/html/marquee/marquee-without-frame-no-crash.html

<script src="../../resources/js-test.js"></script>
foo
<script>
function handler() {
    shouldBeNull('window.open("foo", "bar")');
}
</script>
<template></template>
<iframe onload="handler()"></iframe>
<template>
<marquee>
